WIP: Drafting new OT API
