+++ mpc-aio/src/ot/base/mod.rs
@@ -1,6 +1,6 @@
 pub mod receiver;
 pub mod sender;
 
-pub use super::errors::*;
+use super::OTError;
 pub use receiver::{OTReceive, Receiver};
 pub use sender::{OTSend, Sender};
