+++ mpc-aio/src/ot/extension/mod.rs
@@ -1,7 +1,7 @@
 pub mod receiver;
 pub mod sender;
 
-use super::errors::*;
+use super::OTError;
 
 pub use receiver::{ExtOTReceive, ExtReceiver};
 pub use sender::{ExtOTSend, ExtSender};
